refactor: remove dead packages and dedupe shared code
Over-engineering cleanup (ponytail audit). No behavior change.

Deletions:
- packages/api: unused package, zero consumers (empty DTOs/entities)
- packages/config: 13-string language list folded into @repo/validation;
  dep stripped from all 6 package.json (incl. root)
- apps/web/lib/compose-refs.ts: dead duplicate of the @repo/ui copy
- apps/web/app/dashboard/video-gen: empty dead route
- error-handler: removed unused createErrorResponse; collapsed ApiError/
  mapApiError to {code, retryable} (the only fields anything read)

Dedupes (single source of truth in @repo/validation):
- convertJsonToSrt (was copied in apps/api + apps/web)
- GEMINI_* model constants (was copied in apps/api + packages/workers;
  genai.ts now re-exports so consumers are untouched)
- cn (apps/web/lib/utils.ts now re-exports from @repo/ui)

Minor:
- creator-ai-http: drop dead re-exports, keep creatorAiFetch
- token-manager: unexport file-private helpers
- video-url-tools: switch -> lookup map

Note: config's supportedLanguages renamed to profileLanguages to avoid a
real collision with the existing dubbing supportedLanguages export.

Verified: turbo build 6/6, tests 78/78 (web/api/workers), api type-check
clean, web strict-tsc unchanged at baseline, lint clean.
