+++ apps/api/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './genai';
 export * from './ffmpeg-config';
-export * from './srt-converter';
 export * from './video-url-tools';
