+++ apps/web/lib/creator-ai-http.ts
@@ -1,9 +1,4 @@
-import {
-  CREATOR_AI_USER_AGENT,
-  mergeCreatorAiHeaders,
-} from '@repo/validation';
-
-export { CREATOR_AI_USER_AGENT, mergeCreatorAiHeaders };
+import { mergeCreatorAiHeaders } from '@repo/validation';
 
 export function creatorAiFetch(
   input: RequestInfo | URL,
