--- packages/api/.prettierrc.js
@@ -1,4 +1,0 @@
-/** @type {import("prettier").Config} */
-module.exports = {
-  ...require('@repo/eslint-config/prettier-base'),
-};
