--- packages/api/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@repo/api",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "scripts": {
-    "dev": "pnpm build --watch",
-    "build": "tsc -b -v",
-    "lint": "echo 'No JavaScript files to lint'"
-  },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "./dist/**"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "typesVersions": {
-    "*": {
-      "*": [
-        "src/*"
-      ]
-    }
-  },
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    },
-    "./*": {
-      "import": "./dist/*.js",
-      "require": "./dist/*.js"
-    }
-  },
-  "dependencies": {
-    "@nestjs/mapped-types": "*"
-  },
-  "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "@types/node": "^20.3.1",
-    "ts-loader": "^9.4.3",
-    "ts-node": "^10.9.2",
-    "typescript": "5.5.4"
-  }
-}
