--- packages/api/src/links/dto/create-link.dto.ts
@@ -1,1 +1,0 @@
-export class CreateLinkDto {}
