--- packages/api/src/links/dto/update-link.dto.ts
@@ -1,5 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-
-import { CreateLinkDto } from './create-link.dto';
-
-export class UpdateLinkDto extends PartialType(CreateLinkDto) {}
