--- packages/api/src/links/entities/link.entity.ts
@@ -1,6 +1,0 @@
-export class Link {
-  id: number;
-  url: string;
-  title: string;
-  description: string;
-}
