--- packages/api/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "@repo/typescript-config/nestjs.json",
-  "compilerOptions": {
-    "allowJs": true,
-    "baseUrl": "src",
-    "esModuleInterop": true,
-    "incremental": false,
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "test", "dist", "**/*spec.ts"]
-}
