--- packages/config/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@repo/config",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-   "scripts": {
-    "lint": "echo 'No source JavaScript files to lint'",
-    "build": "tsc -p tsconfig.json"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "eslint": "^9.0.0",
-    "typescript": "^5.5.3"
-  }
-}
