--- packages/config/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2019",
-    "module": "commonjs",
-    "declaration": true,
-    "declarationMap": true,
-    "outDir": "dist",
-    "rootDir": "src",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "lib": ["ES2019", "DOM"]
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
