+++ packages/validations/src/consts/gemini.ts
@@ -1,0 +1,7 @@
+/**
+ * Gemini model ids used across the API and workers. Single source of truth so a
+ * model swap is a one-line change. Override per-environment via env vars if needed.
+ */
+export const GEMINI_TEXT_MODEL = process.env.GEMINI_TEXT_MODEL || 'gemini-3.5-flash';
+export const GEMINI_IMAGE_MODEL = process.env.GEMINI_IMAGE_MODEL || 'gemini-2.5-flash-image';
+export const GEMINI_EMBEDDING_MODEL = process.env.GEMINI_EMBEDDING_MODEL || 'gemini-embedding-001';
