+++ packages/validations/src/consts/index.ts
@@ -2,3 +2,5 @@ export * from './dubbing';
 export * from './credits';
 export * from './ideation';
 export * from './http';
+export * from './gemini';
+export * from './languages';
