+++ packages/workers/src/processor/utils/error-handler.ts
@@ -1,112 +1,47 @@
 export interface ApiError {
   code: string;
-  message: string;
-  details?: string;
-  statusCode: number;
   retryable: boolean;
 }
 
 /**
- * Maps common errors to user-friendly messages and determines if they're retryable
+ * Classifies common errors into a code + whether they're worth retrying.
  */
 export function mapApiError(error: any): ApiError {
   // Network errors
   if (error.code === 'ECONNREFUSED' || error.code === 'ENOTFOUND') {
-    return {
-      code: 'NETWORK_ERROR',
-      message: 'Network error. Please check your internet connection and try again.',
-      details: error.message,
-      statusCode: 500,
-      retryable: true,
-    };
+    return { code: 'NETWORK_ERROR', retryable: true };
   }
 
   // Timeout errors
   if (error.code === 'ECONNABORTED' || error.message?.includes('timeout')) {
-    return {
-      code: 'TIMEOUT_ERROR',
-      message: 'Request timed out. Please try again.',
-      details: error.message,
-      statusCode: 408,
-      retryable: true,
-    };
+    return { code: 'TIMEOUT_ERROR', retryable: true };
   }
 
   // Rate limiting
   if (error.response?.status === 429) {
-    return {
-      code: 'RATE_LIMIT_ERROR',
-      message: 'Too many requests. Please wait a moment and try again.',
-      details: 'Rate limit exceeded',
-      statusCode: 429,
-      retryable: true,
-    };
+    return { code: 'RATE_LIMIT_ERROR', retryable: true };
   }
 
   // YouTube API specific errors
   if (error.response?.status === 403) {
-    return {
-      code: 'FORBIDDEN_ERROR',
-      message: 'Access denied. Please check your YouTube channel permissions.',
-      details: error.response?.data?.error?.message || 'Forbidden',
-      statusCode: 403,
-      retryable: false,
-    };
+    return { code: 'FORBIDDEN_ERROR', retryable: false };
   }
 
   if (error.response?.status === 401) {
-    return {
-      code: 'UNAUTHORIZED_ERROR',
-      message: 'Authentication failed. Please reconnect your YouTube channel.',
-      details: error.response?.data?.error?.message || 'Unauthorized',
-      statusCode: 401,
-      retryable: false,
-    };
+    return { code: 'UNAUTHORIZED_ERROR', retryable: false };
   }
 
   if (error.response?.status === 400) {
-    return {
-      code: 'BAD_REQUEST_ERROR',
-      message: 'Invalid request. Please check your input and try again.',
-      details: error.response?.data?.error?.message || 'Bad request',
-      statusCode: 400,
-      retryable: false,
-    };
+    return { code: 'BAD_REQUEST_ERROR', retryable: false };
   }
 
   // Gemini API errors
   if (error.message?.includes('Gemini') || error.message?.includes('AI')) {
-    return {
-      code: 'AI_SERVICE_ERROR',
-      message: 'AI service temporarily unavailable. Please try again.',
-      details: error.message,
-      statusCode: 500,
-      retryable: true,
-    };
+    return { code: 'AI_SERVICE_ERROR', retryable: true };
   }
 
   // Default error
-  return {
-    code: 'UNKNOWN_ERROR',
-    message: 'An unexpected error occurred. Please try again.',
-    details: error.message || 'Unknown error',
-    statusCode: error.response?.status || 500,
-    retryable: false,
-  };
-}
-
-/**
- * Creates a standardized error response for API endpoints
- */
-export function createErrorResponse(error: any, defaultMessage = 'Internal server error') {
-  const mappedError = mapApiError(error);
-
-  return {
-    error: mappedError.message,
-    code: mappedError.code,
-    details: mappedError.details,
-    retryable: mappedError.retryable,
-  };
+  return { code: 'UNKNOWN_ERROR', retryable: false };
 }
 
 /**
@@ -117,9 +52,6 @@ export function logError(context: string, error: any, additionalInfo?: Record<st
 
   console.error(`[${context}] Error:`, {
     code: mappedError.code,
-    message: mappedError.message,
-    details: mappedError.details,
-    statusCode: mappedError.statusCode,
     retryable: mappedError.retryable,
     originalError: error.message,
     stack: error.stack,
